requirements.txt
# API Interface (Diplomat Module)
fastapi
uvicorn[standard]

# Web Scraping & Parsing
newspaper3k
lxml_html_clean
requests
beautifulsoup4
lxml

# AI & Vector Search
sentence-transformers
faiss-cpu

# AI Summarization
transformers
torch

# Core Utilities
numpy

# Production Grade
loguru
pydantic